fix: resolve issuses with restart for non-admin users
