--- src/shared/utils/formatTimeFromMs.ts
@@ -1,9 +1,0 @@
-export const twoDigits = (num: number) => String(num).padStart(2, "0");
-
-export const formatTimeFromMs = (milliseconds: number): string => {
-  const totalSeconds = Math.floor(milliseconds / 1000);
-  const hours = Math.floor(totalSeconds / 3600);
-  const minutes = Math.floor((totalSeconds % 3600) / 60);
-  const seconds = totalSeconds % 60;
-  return `${twoDigits(hours)}:${twoDigits(minutes)}:${twoDigits(seconds)}`;
-};
